Fix the display of activity during wait
While waiting for a running job, the platform provides the last
lines of log as 'latest solver activity'.  But, depending on the
worker, they are not formatted the same way:
- For a docplex, OPL or CP Optimizer job, these lines need to be
joined with \n.
- For a CPLEX job, these lines must be joined with '' because each
line already ends with \n.

In order to work around this discrepancy, we systematically join
the lines with '\n', and then split the result to remove the empty
lines. What a waste...