Sucessfully used COS
In this version of the code, the file must first be uploaded
manually to Cloud Object Storage.  Then the code creates a
data asset that refers to this file, if it doesn't exist yet.

This saves on upload times if the file was already used, and also
creates much smaller jobs.