Handle the case where a full path is specified
This doesn't really make sense, actually: the specified path or name is just
an id. It's used to set the name of the 'connected data', and to refer to the
object on COS.  But the file is not actually read from disk, and it is therefore
irrelevant whether it actually exists or not on the disk.