Get COS endpoint from the connection
The 'DOWMLClient-connection' that sits in the space has all the information to
connect to COS.  Reading this avoids requiring users to specify the COS endpoint
more than once.

It has the drawback, though, that this endpoint must now be a public one, as the
code using it is now not only in IBM Cloud (as was the case before), but is now
also the DOWML client itself, running on the user's machine.