Keep prompting for commands after Ctrl-C
Previously, the client would stop after a Ctrl-C, except
in 'wait'.  Now, we it never stops.  The point is that hitting
Ctrl-C gets you at the promt, where you can exit cleanly.

It would be nice that hitting Ctrl-C _at the prompt_ would stop
the program, but I don't know how to achieve that.