Prevent spurious empty lines printed between commands
The combination of -i and -c introduced empty lines before each
command that was part of the command line arguments.  This was due
to my assuming that printing an empty string would be invisible.
But it's not: in Python, print() always adds a newline, even when
given an empty string.  So the fix was to not print that newline,
but add it at the end of the program's welcome message.