[Empty] All tests pass after removing all GH references
This is the first commit after the complete rewrite of the
history to remove from commit messages the references to the GH
items that existed in my private repo.  These won't exist in the
new repo, and the commits in the new repo would point nowhere or
to incorrect GH issues.