[#19] Source restructuring to prepare for packaging
The packaging tools (or, probably, the examples I could find
using them) assume that there is an src subdirectory with the
package in it.

Also, move the makefile for the tests and assume that tests are
run from the tests subdirectory. This is what PyCharm does by
default and manually changing this every time is tedious...

I could create more targets into the top-level
Makefile, but I choose not to: the top-level makefile
will be for packaging and distributing.  Testing
is done from the tests directory.