Remove code that was copied from ibm-watson-machine-learning
This function did not accept 'no argument' before. So the code
was copied to dowml and modified to accept not being given an id.
Eventually, the Python API was modified to accept the lack of id.

With the packaging now requiring that the version of the Python
API is high enough, we can remove the copied code.