[Empty] About marking stuff deprecated...
So far, the only deprecated items in the library are attributes
(inline) or function parameters (in get_output). So the usual ways
of using @deprecated decorators will not work, because they only
apply to functions.

Plus they don't allow to actually get rid of the deprecated stuff.
For example, the 'inline' attribute doesn't really exist any more:
it looks like it exists, but actually uses the new 'inputs' attribute
that replaces it.  A decorator would not do that...

But this is something to keep in mind for the future, and here are
links:
- https://deprecation.readthedocs.io/en/latest/
- https://deprecated.readthedocs.io/en/latest/