Add tests to confirm that 'details' works correctly
We had tests that checked we did remove from the details what we
intended to remove, but there were no tests to check that the parts
we expected were still included.  That's why I missed the issue and
thought that WML now dealt correctly with spaces in the filter string.

While at it, move the assert so that it is checked whether we have a
new version of APIClient or an old one.