Use a simpler regexp to catch all outputs
I'm pretty sure that the previous one was not having
the intended effects. What '.*\.*' means is match anything
that ends with a (possibly empty) sequence of dots. This
was probably supposed to be an equivalent of the wildcard
expression '*.*', which should have been '.*\..*'.

But only catching names with a dot is a useless restriction.