Remove useless 'oaas.includeInputData' parameter in payload
The parameter oaas.includeInputData used to govern whether the
input data would be included in the job details after the job
has been processed.  This parameter was obsoleted by the ability
to choose which part of the details one can get (c.f. the code
for get_job_details), and is now ignored by the job processor.