[#49] Save one REST query if possible when creating job
With recent enough versions of the Python client, the method to create a
job can be given the id of the asset that was deployed.  If given this information,
it doesn't have to look for it, and we can therefore save a REST call.

In this implementation, dowml will actually ask for this information itself. But
then the information is cached and all but the first requests are saved.