[#55] Skip trying to delete the WS jod on CPD
If we are connected to a (private) CPD instance (as opposed to
an 'as a Service' instance managed by IBM), then we can't know
the URL of the Watson Studio service (which may even not be
around...).  In this case, we therefore skip trying to find that
URL, as this only introduces noise in the output for no value.